undo a neo4j optimzation that introduced bug
For the get summary endpoint, the likert scale count array was
generated using a cypher "collect" operation instead of in the code.

When a node does not have any values in one of the likert groups the
collect operation skips a value for that group instead of the desired
behavior of returning 0 for that group.